Add a heartbeat, start working on on triggering an election
I'm not sure that the channel is the way to go for triggering
an election. It seems a little unwieldy, since both sides will likely
need to rely on the default.

Maybe it would be better to store the latest hb time as
an attribute of Raft, or set a flag. It could be deleted
before the ticker wait, and checked after. If it hadn't been reset,
then it's election time.