Tests for 2A pass, but this is a mess
At the very least the setToFollowerState fcn needs to be refactored. It
is also handling setting the voter/candidate state. It's also annoying how
many times that function is called. It would be nice to make it more unified
for each of the three states. I'm trying to think of a way to combine all three
states to share a single looped select statement.