I had to take a new approach to avoid some subtle race conditions
With both the worker fcns and the rpc functions sending to the
state change chan, there was the potential for a race condition. E.g.
a state change call from the heartbeat chan right before a
state change due to a follower timeout. The call to the finish
chan in the ticker would deadlock.

The solution was to let the worker functions handle setting the state, and to
let them handle the hb chan receives directly. This also simplified the ticker()
logic, so that the done/finish chan weirdness could be removed.

I think the last thing left is to add a new global chan for indicating when
the RequestVote rpc call recieves a message with a greater term than
the current one. It is subtly different than a heartbeat, but will still
result in a state change for all three worker types.