Fix a bug causing existing candidates vote false in new terms
In the situation when a rf candidate in term n receives a RequestVote
from another candidate term n+1, it should vote for it, and revert to a
follower in term n+1. But the way I had it, the state change was being processed
after the candidate voted. So the candidate remained in term n, and voted
false, because it had already voted for itself in term n.

This commit forces the RPC goroutine to wait for the state change to
complete before deciding its vote.