Don't make a failed leader wait for RPCs to timeout
This can take a long time, the effect of the RPCs can't be changed by
waiting on them. It is best/clearest to forget about them an immediately
transition to a follower.

I made this change after looking at logs where the old leader didn't
participate in subsequen elections after it had failed because
it was blocked waiting for the RPCs to complete.