Prevent a single candidate from continously running campaigns
The way I had it a failed candidate immediately started a new
campaign in the next term. That was my interpretation of this line
in the paper "If election timeout elapses: start new election", but
obviously that doesn't work. If that candidate happens to be
unelectable (as in TestFailAgree2B since it lacks log entries) an
unended series of failed elections occurs.

This commit forces it to wait for another election timeout before
campaigning again, creating a chance for the other instances to
start their own campaign.