Fix bug for TestFaileAgree2B
Just because the entries from old terms can't be committed immediately
doesn't mean they can't be committed at all. Once an entry at the
end of the log, with the current term, can be committed, all of the
previous entries from prior terms need to be committed, in order,
first.