Fix a bug in follower commit index handling
TestRejoin2B exposed a bug in how the commit index was being set in
followers. The implementation assumed that the follower log
would not contain more entries than the leader. So during a test
when an old leader, with a long log, rejoined as a follower some
of its old entries were being committed before they could be rewritten.
The handler now relies on a matching PrevEntry index and term and only
commits once that matches, or any new entries following a sucessful
PrevEntry match.

This change necessitated fixing another bug in the leader side of the
AppendEntries RPC. When there are no entries being send the PrevLogEntry
should be the final entry in the leader log, not the penultimate entry.

TestBackup2B passes now.