Add an index field to each log entry
This is the first step in making log compaction work. The position
of an entry in the rf.logs list won't be useful once logs
are being snapshotted. The next step is to refactor every place in the
code that a reference to the rf.log is used to determine an entry's
index.